requirements.txt
greenlet==3.0.3
numpy==1.26.2
pandas==2.1.4
pyodbc==5.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.9.0
tzdata==2023.3
